Harden canonical decimal boundaries

+++ src/aegis/model/instrument_metadata.cpp
@@ -140,7 +140,7 @@ Result<Quantity> InstrumentMetadata::quantize_quantity(Quantity quantity,
   return remap_numeric_error(quantity.quantize(params_.quantity_step, rounding), "quantity");
 }
 
-Result<Notional> InstrumentMetadata::contract_value(Quantity contracts, std::uint8_t target_scale,
+Result<Notional> InstrumentMetadata::contract_value(Quantity contracts, std::uint64_t target_scale,
                                                     RoundingMode rounding) const {
   const auto aligned = validate_quantity_alignment(contracts);
   if (!aligned) {
